Moved redefined test into separate file
