+++ .github/workflows/testing.yml
@@ -674,6 +674,9 @@ jobs:
       - name: Run Quickstart Test
         run: pytest client/tests/localmode_quickstart_test.py
 
+      - name: Run Redefined Tests
+        run: pytest client/tests/redefined_test.py
+
       - name: Run Register Tests
         run: python client/tests/local_test.py
